requirements.txt
asgiref==3.5.0
black==22.1.0
cffi==1.15.0
click==8.0.4
cryptography==36.0.1
Django==4.0.2
django-rest-knox==4.2.0
djangorestframework==3.13.1
flake8==4.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.1
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
python-dotenv==0.19.2
pytz==2021.3
sqlparse==0.4.2
tomli==2.0.1
typing-extensions==4.1.1
